Dynamically generate list of examples.
